1. []操作考试是否存在扫码或者口令参加呢，如果有那么是否要动态的在批次可参与人和活动可参与人里面增加该考生呢。 []可以做一个组教老师扫考生的id生成的二维码,来临时加入可参与人的这么一个功能.对应理论考试的扫码参与...再多考虑一下,比较好的选择;
  1.
[x]理论考试是如何处理的。是否在两个可参与人里面都增加了呢？

   1.
[x]增加"sp_继教考试批次可参与人_普通增改" ,同时在"参与理论考试"接口里面,更新此可参与人
   2.
[x]同时更新云服务上的sp
 2.
[x]操作考试-继教评分表项目要求,标准分改为一位小数,

  1.
[x]"sp_继教评分表项目要求_增改"中的字段要做相应修改
  2.
[x]"继教操作考试打分情况"中的打分,也要改为一位小数
 3.
[x]参与情况在PC端的哪里显示呢。理论考试是如何体现的呢？是看一个考试还是一个批次的参与人情况呢。

  1.
[x]没有参与，但是可参与的人，其参与情况显示了吗？还是要回头对比下理论考试。
前端做了"可参与人的参与情况"的计算.
而在后台的"参与情况"是没有考虑"可参与但是没有参与"的情况的.
 4.
[]一旦参与了（打分了）就不能再打分了 ,接口"操作考试打分"

  1.
[x]打分是一次性提交[x]没有打分才能打分
  2.
[x]在该批次考试时间内才能打分。但这个考试时间应该是打分老师的操作时间
  3.
[x]建立人可以在PC打分
  4.
[]助教在手机端打分，这就需要有个作为助教可以看到的操作考试，包括其批次，以及批次里面的可参与人。并在满足可打分要求的情况下，能方便的找到考生并打分。打完之后（提交之后）就只能看了。
接口"分页获取某助教可以打分的操作考试活动"
 5.
[x]考生可以看到可以参与的操作考试，
接口:" 分页获取自己可参与的操作考试活动"

  1.
[x]可以看到评分表 ,接口"获取某操作考试评分表信息"
  2.
[x]可以看到自己的参与情况，包括详细的评分。 接口"获取操作考试打分情况"和"获取某人的操作考试打分情况",这两个接口,和上面的"获取某操作考试评分表信息"结合起来,可以说明一个考试的操作考试情况;

+++ ScientificResearch/Models/Table/继教评分表项目要求.cs
@@ -33,7 +33,7 @@ namespace ScientificResearch.Models
 		/// 
 		/// </summary>
         [Required(ErrorMessage ="请提供[标准分]")]
-		public int 标准分 { get; set; }
+		public decimal 标准分 { get; set; }
         /// <summary>
 		/// 
 		/// </summary>
