+++ ScientificResearch/Models/View/v_继教操作考试评分表.cs
@@ -43,7 +43,7 @@ namespace ScientificResearch.Models
         /// <summary>
 		/// 
 		/// </summary>
-        public int? 总分值 { get; set; }
+        public decimal? 总分值 { get; set; }
         /// <summary>
 		/// 
 		/// </summary>
